+++ R/util_classify.R
@@ -1,48 +1,61 @@
 #' util_classify
 #'
-#' @description Classify a raster into proportions based upon a vector of class
-#' weightings.
+#' @description  Classify continuous landscapes into landscapes with discrete classes
 #'
-#' @details  The number of elements in the weighting vector determines the number of classes
-#' in the resulting matrix. The classes start with the value 0.
+#' @details
+#' Mode 1: Convienient interface to classInt::classIntervals to classify landscapes.
+#'
+#' Mode 2: The number of elements in the weighting vector determines the number of classes
+#' in the resulting matrix. The classes start with the value 1.
 #' If non-numerical levels are required, the user can specify a vector to turn the
 #' numerical factors into other data types, for example into character strings (i.e. class labels).
 #' If the numerical vector of weightings does not sum up to 1, the sum of the
 #' weightings is divided by the number of elements in the weightings vector and this is then used for the classification.
 #'
-#' For a given 'real' landscape the number of classes and the weightings are
-#' extracted and used to classify the given nlm landscape (any given weighting parameter is
+#' Mode 3: For a given 'real' landscape the number of classes and the weightings are
+#' extracted and used to classify the given landscape (any given weighting parameter is
 #' overwritten in this case!). If an optional mask value is given the corresponding
-#' class from the 'real' landscape is cut from the nlm landscape beforehand.
+#' class from the 'real' landscape is cut from the landscape beforehand.
 #'
-#' @param x 2D matrix
-#' @param weighting Vector of numeric values.
+#' @param x raster
+#' @param n Number of classes
+#' @param style Style of breaks (see classInt::classInvervals() for more details)
+#' @param weighting Vector of numeric values that are considered to be habitat percentages (see details)
 #' @param level_names Vector of names for the factor levels.
-#' @param real_land Raster with real landscape
+#' @param real_land Raster with real landscape (see details)
 #' @param mask_val Value to mask (refers to real_land)
 #'
 #' @return RasterLayer
 #'
 #' @examples
-#' weight <- c(0.5, 0.25, 0.25)
-#' util_classify(fbmmap, weight,
+#' # Mode 1
+#' util_classify(fbmmap,
+#'               n = 3,
+#'               style = "fisher",
 #'               level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))
 #'
+#' # Mode 2
+#' util_classify(fbmmap,
+#'               weighting = c(0.5, 0.25, 0.25),
+#'               level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))
+#'
+#' # Mode 3
+#' real_land <- util_classify(grdmap,
+#'               n = 3,
+#'               level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))
+#'
+#' fbmmap_real <- util_classify(fbmmap, real_land = real_land)
+#' fbmmap_mask <- util_classify(fbmmap, real_land = real_land, mask_val = 1)
+#'
 #' \dontrun{
-#' rland <- util_classify(NLMR::nlm_planargradient(200,200),
-#'                          c(.4,.2,.4),
-#'                          c("Land use 1", "Water", "Land use 2"))
-#'
-#' resu <- util_classify(fbmmap, real_land = rland)
-#' resu_mask <- util_classify(fbmmap, real_land = rland, mask_val = 1)
-#'
-#' visu <- list(
+#' landscapes <- list(
 #' '1 nlm' = fbmmap,
-#' '2 real' = rland,
-#' '3 result' = resu,
-#' '4 result with mask' = resu_mask
+#' '2 real' = real_land,
+#' '3 result' = fbmmap_real,
+#' '4 result with mask' = fbmmap_mask
 #' )
-#' util_facetplot(visu)
+#'
+#' util_facetplot(landscapes)
 #' }
 #'
 #' @aliases util_classify
@@ -50,15 +63,19 @@
 #'
 #' @export
 #'
-
-util_classify <- function(x, weighting, level_names = NULL,
-                          real_land = NULL, mask_val = NULL) {
+util_classify <- function(x,
+                          n,
+                          style = "fisher",
+                          weighting = NULL,
+                          level_names = NULL,
+                          real_land = NULL,
+                          mask_val = NULL) {
 
   # Check function arguments ----
   checkmate::assert_class(x, "RasterLayer")
-  if (is.null(real_land)) {
-      checkmate::assert_numeric(weighting)
-  }else{
+
+  # Classify based on real landscape ----
+  if (!is.null(real_land)) {
       checkmate::assert_class(real_land, "RasterLayer")
       frq <- tibble::as_tibble(raster::freq(real_land))
       if (!is.null(mask_val)) {
@@ -66,17 +83,24 @@ util_classify <- function(x, weighting, level_names = NULL,
           x <- raster::mask(x, real_land, maskvalue = mask_val)
       }
       weighting <- frq$count / sum(frq$count)
+
+      x <- .classify(x, weighting)
+
   }
 
-  # Calculate cum. proportions and boundary values ----
-  cumulative_proportions <- util_w2cp(weighting)
-  boundary_values <- util_calc_boundaries(raster::values(x),
-                                          cumulative_proportions)
+  if(is.null(real_land)){
 
-  # Classify the matrix based on the boundary values ----
-  raster::values(x) <- findInterval(raster::values(x),
-                                    boundary_values,
-                                    rightmost.closed = TRUE)
+
+    if (is.null(weighting)){
+      breaks <- classInt::classIntervals(raster::getValues(x), n = n, style= style)
+      x <-  raster::cut(x, breaks=breaks$brks, include.lowest=T)
+    }
+
+    if (!is.null(weighting)){
+      x <- .classify(x, weighting)
+    }
+
+  }
 
   # If level_names are not NULL, add them as specified ----
   if (!is.null(level_names)) {
@@ -90,4 +114,24 @@ util_classify <- function(x, weighting, level_names = NULL,
   }
 
   return(x)
+}
+
+
+.classify <- function(x, weighting){
+
+  # Calculate cum. proportions and boundary values ----
+  cumulative_proportions <- util_w2cp(weighting)
+  boundary_values <- util_calc_boundaries(raster::values(x),
+                                          cumulative_proportions)
+
+  # Classify the matrix based on the boundary values ----
+  raster::values(x) <- findInterval(raster::values(x),
+                                    boundary_values,
+                                    rightmost.closed = TRUE)
+
+  # first class should be 1
+  x <- x + 1
+
+  return(x)
+
 }